+++ packages/core-js/internals/define-built-in.js
@@ -1,5 +1,5 @@
 var isCallable = require('../internals/is-callable');
-var createNonEnumerableProperty = require('../internals/create-non-enumerable-property');
+var definePropertyModule = require('../internals/object-define-property');
 var makeBuiltIn = require('../internals/make-built-in');
 var defineGlobalProperty = require('../internals/define-global-property');
 
@@ -15,6 +15,11 @@ module.exports = function (O, key, value, options) {
     if (!options.unsafe) delete O[key];
     else if (O[key]) simple = true;
     if (simple) O[key] = value;
-    else createNonEnumerableProperty(O, key, value);
+    else definePropertyModule.f(O, key, {
+      value: value,
+      enumerable: false,
+      configurable: !options.nonConfigurable,
+      writable: !options.nonWritable
+    });
   } return O;
 };