+++ packages/core-js/modules/es.number.epsilon.js
@@ -2,6 +2,6 @@ var $ = require('../internals/export');
 
 // `Number.EPSILON` constant
 // https://tc39.es/ecma262/#sec-number.epsilon
-$({ target: 'Number', stat: true }, {
+$({ target: 'Number', stat: true, nonConfigurable: true, nonWritable: true }, {
   EPSILON: Math.pow(2, -52)
 });