+++ packages/core-js/modules/esnext.math.deg-per-rad.js
@@ -2,6 +2,6 @@ var $ = require('../internals/export');
 
 // `Math.DEG_PER_RAD` constant
 // https://rwaldron.github.io/proposal-math-extensions/
-$({ target: 'Math', stat: true, forced: true }, {
+$({ target: 'Math', stat: true, nonConfigurable: true, nonWritable: true, forced: true }, {
   DEG_PER_RAD: Math.PI / 180
 });